--- apps/raddix/src/components/api-table/index.ts
@@ -1,1 +1,0 @@
-export * from './api-table';
