--- apps/raddix/src/components/button/index.tsx
@@ -1,1 +1,0 @@
-export * from './button.comp';
