--- apps/raddix/src/components/header/index.tsx
@@ -1,1 +1,0 @@
-export * from './header.comp';
