--- apps/raddix/src/components/hero/index.tsx
@@ -1,1 +1,0 @@
-export * from './hero.comp';
