--- apps/raddix/src/components/language/index.ts
@@ -1,1 +1,0 @@
-export * from './language';
