--- apps/raddix/src/components/menu-mobile/index.ts
@@ -1,1 +1,0 @@
-export * from './menu-mobile';
