--- apps/raddix/src/components/menu/index.tsx
@@ -1,1 +1,0 @@
-export * from './menu.comp';
