--- apps/raddix/src/components/message/index.ts
@@ -1,1 +1,0 @@
-export * from './message';
