+++ apps/raddix/src/components/search-bar.tsx
