--- apps/raddix/src/components/search-bar/index.ts
@@ -1,1 +1,0 @@
-export * from './search-bar';
