--- apps/raddix/src/components/search-menu/index.ts
@@ -1,1 +1,0 @@
-export * from './search-menu';
