--- apps/raddix/src/components/styling/index.ts
@@ -1,1 +1,0 @@
-export * from './styling';
