--- apps/raddix/src/layouts/app/index.tsx
@@ -1,1 +1,0 @@
-export { default as AppLayout } from './app.layout';
