--- apps/raddix/src/layouts/docs/index.tsx
@@ -1,1 +1,0 @@
-export * from './docs.layout';
