--- apps/raddix/src/layouts/home/index.tsx
@@ -1,1 +1,0 @@
-export * from './home.layout';
